+++ src/common-helpers.ts
@@ -1,0 +1,21 @@
+import { z } from "zod";
+
+export const tryToTransform = <T>(
+  schema: z.ZodEffects<z.ZodTypeAny, T>,
+  sample: T,
+) => {
+  try {
+    return typeof schema.parse(sample);
+  } catch (e) {
+    return undefined;
+  }
+};
+
+/**
+ * @desc isNullable() and isOptional() validate the schema's input
+ * @desc They always return true in case of coercion, which should be taken into account when depicting response
+ */
+export const hasCoercion = (schema: z.ZodTypeAny): boolean =>
+  "coerce" in schema._def && typeof schema._def.coerce === "boolean"
+    ? schema._def.coerce
+    : false;