+++ src/integration-helpers.spec.ts
@@ -1,27 +1,8 @@
 import { describe, expect, test } from "vitest";
 import { z } from "zod";
-import {
-  hasCoercion,
-  lcFirst,
-  makeCleanId,
-  makeEventFnSchema,
-} from "./integration-helpers";
+import { lcFirst, makeCleanId, makeEventFnSchema } from "./integration-helpers";
 
 describe("Integration helpers", () => {
-  describe("hasCoercion()", () => {
-    test.each([
-      { schema: z.string(), coercion: false },
-      { schema: z.coerce.string(), coercion: true },
-      { schema: z.boolean({ coerce: true }), coercion: true },
-      { schema: z.custom(), coercion: false },
-    ])(
-      "should check the presence and value of coerce prop %#",
-      ({ schema, coercion }) => {
-        expect(hasCoercion(schema)).toBe(coercion);
-      },
-    );
-  });
-
   describe("makeCleanId()", () => {
     test.each([
       ["get"],