+++ src/zts.ts
@@ -1,12 +1,7 @@
 import ts from "typescript";
 import { z } from "zod";
-
-import {
-  hasCoercion,
-  lcFirst,
-  makeCleanId,
-  tryToTransform,
-} from "./integration-helpers";
+import { hasCoercion, tryToTransform } from "./common-helpers";
+import { lcFirst, makeCleanId } from "./integration-helpers";
 import { HandlingRules, walkSchema } from "./schema-walker";
 import {
   LiteralType,