+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 ## Version 0
 
+### v0.11.0
+
+- Featuring `Documentation` class:
+  - Ability to generate the documentation of your Socket.IO-based application according to AsyncAPI standard;
+  - Using a custom protocol `socket.io` that extends WebSockets bindings for describing acknowledgements and handshake;
+  - Compliance with AsyncAPI version 2.5.0 so far (will be increased later);
+  - Following features are not supported yet:
+    - Examples,
+    - `z.lazy()` and handling of circular references,
+    - References and component-based composition of the document,
+    - Informative errors.
+  - Since AsyncAPI does not yet support `prefixItems` feature for describing tuples, those are depicted as objects
+    having numeric properties. I found it acceptable at the moment because
+    [Arrays are Objects](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array);
+  - See the example of the generated documentation [here](example/example-documentation.yaml).
+
+```typescript
+import { Documentation } from "zod-sockets";
+
+const yamlString = new Documentation({
+  config,
+  actions,
+  version: "1.2.3",
+  title: "Example APP",
+  servers: { example: { url: "https://example.com/socket.io" } },
+}).getSpecAsYaml();
+```
+
 ### v0.10.0
 
 - Important changes to configuration: